fix: 초기화면 블러효과 적용
