+++ src/css/modal.style.css
@@ -1,6 +1,5 @@
 .modal-overlay {
     box-sizing: border-box;
-    display: none;
     position: fixed;
     top: 0;
     left: 0;
